Split code to modules
